% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.edbl_table}
\alias{autoplot.edbl_table}
\title{Auto plot or \code{ggplot2} of an edibble design}
\usage{
\method{autoplot}{edbl_table}(
  .edibble,
  aspect_ratio = 1,
  shape = "circle",
  text = FALSE,
  image = NULL,
  trts = NULL,
  scales = NULL
)
}
\arguments{
\item{.edibble}{An edibble desgin, an edibble table or an edibble graph.}

\item{...}{Unused at the moment.}

\item{title}{The title of the plot. By default it is the
name of the edibble design.}
}
\value{
A \code{ggplot} object.
}
\description{
Auto plot or \code{ggplot2} of an edibble design
}
