% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deggust-package.R
\docType{package}
\name{deggust-package}
\alias{deggust}
\alias{deggust-package}
\title{deggust: Visualise experimental designs from edibble as ggplot graphics}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Visualisation of designs as ggplot graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://deggust.emitanaka.org}
  \item \url{https://github.com/emitanaka/deggust}
  \item Report bugs at \url{https://github.com/emitanaka/deggust/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
