% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_aes_select}
\alias{scale_aes_select}
\title{Change selected scale where multiple scales of same aesthetic exist}
\usage{
scale_aes_select(i, f, ...)
}
\arguments{
\item{i}{An integer denoting the scale number.}

\item{f}{The scale function or the scale object.}

\item{...}{The arguments to the scale function \code{f}.}
}
\description{
Change selected scale where multiple scales of same aesthetic exist
}
\examples{
scale_aes_select(1, ggplot2::scale_fill_manual, values = c("red", "white", "blue"))
scale_aes_select(1, ggplot2::scale_fill_manual(values = c("red", "white", "blue")))

}
\seealso{
scale_fills
}
