% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fills}
\alias{scale_fills}
\alias{scale_fill1_binned}
\alias{scale_fill2_binned}
\alias{scale_fill3_binned}
\alias{scale_fill1_brewer}
\alias{scale_fill2_brewer}
\alias{scale_fill3_brewer}
\alias{scale_fill1_continuous}
\alias{scale_fill2_continuous}
\alias{scale_fill3_continuous}
\alias{scale_fill1_date}
\alias{scale_fill2_date}
\alias{scale_fill3_date}
\alias{scale_fill1_datetime}
\alias{scale_fill2_datetime}
\alias{scale_fill3_datetime}
\alias{scale_fill1_discrete}
\alias{scale_fill2_discrete}
\alias{scale_fill3_discrete}
\alias{scale_fill1_manual}
\alias{scale_fill2_manual}
\alias{scale_fill3_manual}
\title{Colour scales for multiple scales of the same aesthetic}
\usage{
scale_fill1_binned(...)

scale_fill2_binned(...)

scale_fill3_binned(...)

scale_fill1_brewer(...)

scale_fill2_brewer(...)

scale_fill3_brewer(...)

scale_fill1_continuous(...)

scale_fill2_continuous(...)

scale_fill3_continuous(...)

scale_fill1_date(...)

scale_fill2_date(...)

scale_fill3_date(...)

scale_fill1_datetime(...)

scale_fill2_datetime(...)

scale_fill3_datetime(...)

scale_fill1_discrete(...)

scale_fill2_discrete(...)

scale_fill3_discrete(...)

scale_fill1_manual(...)

scale_fill2_manual(...)

scale_fill3_manual(...)
}
\arguments{
\item{...}{The arguments for the corresponding scale function.}
}
\description{
This function has the same argument as the corresponding \code{ggplot2} scale
functions where the index after the aesthetic name is omitted from the function.
E.g. \code{scale_fill1_binned} has the same argument as \code{scale_filled_binned}.
Where there is multiple \code{fill} scales, the index after the aesthetic name
determines which \code{fill} scale is modified.
}
\seealso{
scale_aes_select
}
