% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.edibble}
\alias{autoplot.edibble}
\alias{autoplot.EdibbleDesign}
\alias{autoplot.edbl_graph}
\alias{autoplot.edbl_table}
\title{Auto plot or \code{ggplot2} of an edibble design}
\usage{
\method{autoplot}{EdibbleDesign}(.edibble, view = c("high", "low"), ..., main = NULL)

\method{autoplot}{edbl_graph}(.edibble, view = c("high", "low"), ..., main = NULL)

\method{autoplot}{edbl_table}(.edibble, view = c("high", "low"), ..., main = NULL)
}
\arguments{
\item{.edibble}{An edibble desgin, an edibble table or an edibble graph.}

\item{view}{A string specifying the level of the view ("high" or "low") to plot.
By default it is "high".}

\item{...}{Additional parameters passed to \code{plot.igraph}.}

\item{main}{The title of the plot. By default it is the
name of the edibble design.}
}
\value{
A \code{ggraph} and \code{ggplot} object.
}
\description{
This plot produces the same type of plot using the \code{plot} functions in
\code{edibble} package except the plots are \code{ggplot}
}
\seealso{
See \code{\link[edibble]{plot.edibble}} for the base plot version of this plot.
}
